backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
requests>=2.31
python-dotenv>=1.0
gunicorn>=21.0
whitenoise>=6.0
dj-database-url>=2.1
psycopg2-binary>=2.9
